+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/blockmanagement/ReplicaUnderConstruction.java
@@ -76,11 +76,11 @@ public class ReplicaUnderConstruction extends Replica {
   private long generationStamp;
 
   public ReplicaUnderConstruction(ReplicaState state, int storageId,
-      long blockId, int inodeId, int bucketId, long genStamp) {
+      long blockId, long inodeId, int bucketId, long genStamp) {
     this(state, storageId, blockId, inodeId, bucketId, false, genStamp);
   }
   public ReplicaUnderConstruction(ReplicaState state, int storageId,
-      long blockId, int inodeId, int bucketId, boolean chosenAsPrimary, long generationStamp) {
+      long blockId, long inodeId, int bucketId, boolean chosenAsPrimary, long generationStamp) {
     super(storageId, blockId, inodeId, bucketId);
     this.state = state;
     this.chosenAsPrimary = chosenAsPrimary;
